refactor: adapt formula to new core version
