+++ smartwatts/__main__.py
@@ -33,23 +33,22 @@ import sys
 from typing import Dict
 
 from powerapi import __version__ as powerapi_version
+from powerapi.backend_supervisor import BackendSupervisor
 from powerapi.dispatcher import RouteTable
 
 from powerapi.cli import ConfigValidator
 from powerapi.cli.tools import store_true, CommonCLIParser
 from powerapi.cli.generator import ReportModifierGenerator, PullerGenerator, PusherGenerator
-from powerapi.message import DispatcherStartMessage
+from powerapi.exception import PowerAPIException
 from powerapi.report import HWPCReport, PowerReport
 from powerapi.dispatch_rule import HWPCDispatchRule, HWPCDepthLevel
 from powerapi.filter import Filter
-from powerapi.actor import InitializationException, PowerAPIException
-from powerapi.supervisor import Supervisor
+from powerapi.actor import InitializationException
 
 from smartwatts import __version__ as smartwatts_version
 from smartwatts.report import FormulaReport
-from smartwatts.dispatcher import SmartwattsDispatcherActor
-from smartwatts.actor import SmartWattsFormulaActor, SmartwattsValues
-from smartwatts.context import SmartWattsFormulaScope, SmartWattsFormulaConfig
+from smartwatts.smartwatts_dispatcher_actor import SmartWattsDispatcherActor
+from smartwatts.smartwatts_formula_actor import SmartWattsFormulaActor, SmartWattsFormulaScope, SmartWattsFormulaConfig
 from smartwatts.topology import CPUTopology
 
 
@@ -61,30 +60,41 @@ def generate_smartwatts_parser():
     parser = CommonCLIParser()
 
     # Formula control parameters
-    parser.add_argument('disable-cpu-formula', help='Disable CPU formula', flag=True, type=bool, default=False, action=store_true)
-    parser.add_argument('disable-dram-formula', help='Disable DRAM formula', flag=True, type=bool, default=False, action=store_true)
+    parser.add_argument('disable-cpu-formula', help='Disable CPU formula', flag=True, type=bool, default=False,
+                        action=store_true)
+    parser.add_argument('disable-dram-formula', help='Disable DRAM formula', flag=True, type=bool, default=False,
+                        action=store_true)
 
     # Formula RAPL reference event
-    parser.add_argument('cpu-rapl-ref-event', help='RAPL event used as reference for the CPU power models', default='RAPL_ENERGY_PKG')
-    parser.add_argument('dram-rapl-ref-event', help='RAPL event used as reference for the DRAM power models', default='RAPL_ENERGY_DRAM')
+    parser.add_argument('cpu-rapl-ref-event', help='RAPL event used as reference for the CPU power models',
+                        default='RAPL_ENERGY_PKG')
+    parser.add_argument('dram-rapl-ref-event', help='RAPL event used as reference for the DRAM power models',
+                        default='RAPL_ENERGY_DRAM')
 
     # CPU topology information
     parser.add_argument('cpu-tdp', help='CPU TDP (in Watt)', type=int, default=125)
     parser.add_argument('cpu-base-clock', help='CPU base clock (in MHz)', type=int, default=100)
     parser.add_argument('cpu-frequency-min', help='CPU minimal frequency (in MHz)', type=int, default=100)
     parser.add_argument('cpu-frequency-base', help='CPU base frequency (in MHz)', type=int, default=2300)
-    parser.add_argument('cpu-frequency-max', help='CPU maximal frequency (In MHz, with Turbo-Boost)', type=int, default=4000)
+    parser.add_argument('cpu-frequency-max', help='CPU maximal frequency (In MHz, with Turbo-Boost)', type=int,
+                        default=4000)
 
     # Formula error threshold
-    parser.add_argument('cpu-error-threshold', help='Error threshold for the CPU power models (in Watt)', type=float, default=2.0)
-    parser.add_argument('dram-error-threshold', help='Error threshold for the DRAM power models (in Watt)', type=float, default=2.0)
+    parser.add_argument('cpu-error-threshold', help='Error threshold for the CPU power models (in Watt)', type=float,
+                        default=2.0)
+    parser.add_argument('dram-error-threshold', help='Error threshold for the DRAM power models (in Watt)', type=float,
+                        default=2.0)
 
     # Sensor information
-    parser.add_argument('sensor-report-sampling-interval', help='The frequency with which measurements are made (in milliseconds)', type=int, default=1000)
+    parser.add_argument('sensor-report-sampling-interval',
+                        help='The frequency with which measurements are made (in milliseconds)', type=int, default=1000)
 
     # Learning parameters
-    parser.add_argument('learn-min-samples-required', help='Minimum amount of samples required before trying to learn a power model', type=int, default=10)
-    parser.add_argument('learn-history-window-size', help='Size of the history window used to keep samples to learn from', type=int, default=60)
+    parser.add_argument('learn-min-samples-required',
+                        help='Minimum amount of samples required before trying to learn a power model', type=int,
+                        default=10)
+    parser.add_argument('learn-history-window-size',
+                        help='Size of the history window used to keep samples to learn from', type=int, default=60)
     parser.add_argument('real-time-mode', help='Pass the wait for reports from 4 ticks to 1', type=bool, default=False)
     return parser
 
@@ -96,83 +106,143 @@ def filter_rule(_):
     return True
 
 
-def setup_cpu_formula_actor(supervisor, fconf, route_table, report_filter, cpu_topology, formula_pushers, power_pushers):
+def setup_formula_actor(supervisor: BackendSupervisor, global_configuration: Dict, route_table: RouteTable,
+                        report_filter: Filter, cpu_topology: CPUTopology, formula_pushers: Dict,
+                        power_pushers: Dict, rapl_event: str, error_threshold: int, dispatcher_name: str,
+                        device_id: str):
     """
     Setup CPU formula actor.
     :param supervisor: Actor supervisor
-    :param fconf: Global configuration
+    :param global_configuration: Global configuration
     :param route_table: Reports routing table
     :param report_filter: Reports filter
     :param cpu_topology: CPU topology information
-    :param pushers: Reports pushers
+    :param power_pushers: Power Reports pushers
+    :param formula_pushers: Formula Reports pushers
+    :param rapl_event: RAPL event associated with the Dispatcher
+    :param error_threshold: Error threshold related to Formulas managed by the Dispatcher
+    :param dispatcher_name: Dispatcher Name
+    :param device_id: Id of the device related to the Dispatcher
     """
-    formula_config = SmartWattsFormulaConfig(SmartWattsFormulaScope.CPU, fconf['sensor-report-sampling-interval'],
-                                             fconf['cpu-rapl-ref-event'], fconf['cpu-error-threshold'],
-                                             cpu_topology, fconf['learn-min-samples-required'],
-                                             fconf['learn-history-window-size'], fconf['real-time-mode'])
-    dispatcher_start_message = DispatcherStartMessage('system', 'cpu_dispatcher', SmartWattsFormulaActor,
-                                                      SmartwattsValues(formula_pushers, power_pushers,
-                                                                       formula_config), route_table, 'cpu')
-    cpu_dispatcher = supervisor.launch(SmartwattsDispatcherActor, dispatcher_start_message)
-    report_filter.filter(filter_rule, cpu_dispatcher)
+    formula_config = SmartWattsFormulaConfig(scope=SmartWattsFormulaScope.CPU,
+                                             reports_frequency=global_configuration['sensor-report-sampling-interval'],
+                                             rapl_event=rapl_event,
+                                             error_threshold=error_threshold,
+                                             cpu_topology=cpu_topology,
+                                             min_samples_required=global_configuration['learn-min-samples-required'],
+                                             history_window_size=global_configuration['learn-history-window-size'],
+                                             real_time_mode=global_configuration['real-time-mode'])
+
+    dispatcher = SmartWattsDispatcherActor(name=dispatcher_name, formula_init_function=SmartWattsFormulaActor,
+                                           power_pushers=power_pushers, formula_pushers=formula_pushers,
+                                           route_table=route_table, device_id=device_id,
+                                           formula_config=formula_config)
+
+    supervisor.launch_actor(dispatcher)
+
+    report_filter.filter(filter_rule, dispatcher)
 
 
-def setup_dram_formula_actor(supervisor, fconf, route_table, report_filter, cpu_topology, formula_pushers, power_pushers):
-    """
-    Setup DRAM formula actor.
-    :param supervisor: Actor supervisor
-    :param fconf: Global configuration
-    :param route_table: Reports routing table
-    :param report_filter: Reports filter
-    :param cpu_topology: CPU topology information
-    :param pushers: Reports pushers
-    :return: Initialized DRAM dispatcher actor
-    """
-    formula_config = SmartWattsFormulaConfig(SmartWattsFormulaScope.DRAM,
-                                             fconf['sensor-report-sampling-interval'],
-                                             fconf['dram-rapl-ref-event'],
-                                             fconf['dram-error-threshold'],
-                                             cpu_topology,
-                                             fconf['learn-min-samples-required'],
-                                             fconf['learn-history-window-size'],
-                                             fconf['real-time-mode'])
-    dispatcher_start_message = DispatcherStartMessage('system',
-                                                      'dram_dispatcher',
-                                                      SmartWattsFormulaActor,
-                                                      SmartwattsValues(formula_pushers,
-                                                                       power_pushers, formula_config),
-                                                      route_table, 'dram')
-    dram_dispatcher = supervisor.launch(SmartwattsDispatcherActor, dispatcher_start_message)
-    report_filter.filter(filter_rule, dram_dispatcher)
+# def setup_cpu_formula_actor(supervisor, global_configuration, route_table, report_filter, cpu_topology,
+# formula_pushers,
+#                             power_pushers):
+#     """
+#     Setup CPU formula actor.
+#     :param supervisor: Actor supervisor
+#     :param global_configuration: Global configuration
+#     :param route_table: Reports routing table
+#     :param report_filter: Reports filter
+#     :param cpu_topology: CPU topology information
+#     :param power_pushers: Power Reports pushers
+#     :param formula_pushers: Formula Reports pushers
+#     """
+#     formula_config = SmartWattsFormulaConfig(scope=SmartWattsFormulaScope.CPU,
+#                                              reports_frequency=global_configuration['sensor-report-sampling-interval'],
+#                                              rapl_event=global_configuration['cpu-rapl-ref-event'],
+#                                              error_threshold=global_configuration['cpu-error-threshold'],
+#                                              cpu_topology=cpu_topology,
+#                                              min_samples_required=global_configuration['learn-min-samples-required'],
+#                                              history_window_size=global_configuration['learn-history-window-size'],
+#                                              real_time_mode=global_configuration['real-time-mode'])
+#
+#     cpu_dispatcher = SmartWattsDispatcherActor(name='cpu_dispatcher', formula_init_function=SmartWattsFormulaActor,
+#                                                power_pushers=power_pushers, formula_pushers=formula_pushers,
+#                                                route_table=route_table, device_id='cpu',
+#                                                formula_config=formula_config)
+#
+#     # dispatcher_start_message = DispatcherStartMessage('system', 'cpu_dispatcher', SmartWattsFormulaActor,
+#     #                                                  SmartwattsValues(formula_pushers, power_pushers,
+#     #                                                                   formula_config), route_table, 'cpu')
+#     # cpu_dispatcher = supervisor.launch(SmartwattsDispatcherActor, dispatcher_start_message)
+#     supervisor.launch_actor(cpu_dispatcher)
+#
+#     report_filter.filter(filter_rule, cpu_dispatcher)
 
 
-def run_smartwatts(args) -> None:
+# def setup_dram_formula_actor(supervisor, global_configuration, route_table, report_filter, cpu_topology,
+#                              formula_pushers,
+#                              power_pushers):
+#     """
+#     Setup DRAM formula actor.
+#     :param supervisor: Actor supervisor
+#     :param global_configuration: Global configuration
+#     :param route_table: Reports routing table
+#     :param report_filter: Reports filter
+#     :param cpu_topology: CPU topology information
+#     :param power_pushers: Power Reports pushers
+#     :param formula_pushers: Formula Reports pushers
+#     :return: Initialized DRAM dispatcher actor
+#     """
+#     formula_config = SmartWattsFormulaConfig(SmartWattsFormulaScope.DRAM,
+#                                              global_configuration['sensor-report-sampling-interval'],
+#                                              global_configuration['dram-rapl-ref-event'],
+#                                              global_configuration['dram-error-threshold'],
+#                                              cpu_topology,
+#                                              global_configuration['learn-min-samples-required'],
+#                                              global_configuration['learn-history-window-size'],
+#                                              global_configuration['real-time-mode'])
+#     dispatcher_start_message = DispatcherStartMessage('system',
+#                                                       'dram_dispatcher',
+#                                                       SmartWattsFormulaActor,
+#                                                       SmartwattsValues(formula_pushers,
+#                                                                        power_pushers, formula_config),
+#                                                       route_table, 'dram')
+#     dram_dispatcher = supervisor.launch(SmartwattsDispatcherActor, dispatcher_start_message)
+#     report_filter.filter(filter_rule, dram_dispatcher)
+
+
+def run_smartwatts(args) -> BackendSupervisor:
     """
     Run PowerAPI with the SmartWatts formula.
     :param args: CLI arguments namespace
     :param logger: Logger to use for the actors
     """
-    fconf = args
+    global_configuration = args
 
     logging.info('SmartWatts version %s using PowerAPI version %s', smartwatts_version, powerapi_version)
 
-    if fconf['disable-cpu-formula'] and fconf['disable-dram-formula']:
+    if global_configuration['disable-cpu-formula'] and global_configuration['disable-dram-formula']:
         logging.error('You need to enable at least one formula')
         return
 
     route_table = RouteTable()
     route_table.dispatch_rule(HWPCReport, HWPCDispatchRule(HWPCDepthLevel.SOCKET, primary=True))
 
-    cpu_topology = CPUTopology(fconf['cpu-tdp'], fconf['cpu-base-clock'], fconf['cpu-frequency-min'], fconf['cpu-frequency-base'], fconf['cpu-frequency-max'])
+    cpu_topology = CPUTopology(tdp=global_configuration['cpu-tdp'],
+                               freq_bclk=global_configuration['cpu-base-clock'],
+                               ratio_min=global_configuration['cpu-frequency-min'],
+                               ratio_base=global_configuration['cpu-frequency-base'],
+                               ratio_max=global_configuration['cpu-frequency-max'])
 
     report_filter = Filter()
 
-    report_modifier_list = ReportModifierGenerator().generate(fconf)
+    report_modifier_list = ReportModifierGenerator().generate(global_configuration)
 
-    supervisor = Supervisor(args['verbose'], args['actor_system'])
+    supervisor = BackendSupervisor(global_configuration['stream'])  # Supervisor(args['verbose'], args['actor_system'])
 
     def term_handler(_, __):
-        supervisor.shutdown()
+        logging.info('SmartWatts is shutting down...')
+        supervisor.kill_actors()
         sys.exit(0)
 
     signal.signal(signal.SIGTERM, term_handler)
@@ -181,53 +251,75 @@ def run_smartwatts(args) -> None:
         logging.info('Starting SmartWatts actors...')
 
         pusher_generator = PusherGenerator()
-        pusher_generator.add_model_factory('FormulaReport', FormulaReport)
-        pushers_info = pusher_generator.generate(args)
-        pushers_formula = {}
-        pushers_power = {}
+        pusher_generator.add_report_class('FormulaReport', FormulaReport)
+        pushers_info = pusher_generator.generate(global_configuration)
+        formula_pushers = {}
+        power_pushers = {}
 
         for pusher_name in pushers_info:
-            pusher_cls, pusher_start_message = pushers_info[pusher_name]
+            print('Creating Pusher '+pusher_name)
+            pusher = pushers_info[pusher_name]
 
-            if pusher_start_message.database.report_type == PowerReport:
-                pushers_power[pusher_name] = supervisor.launch(pusher_cls, pusher_start_message)
-            elif pusher_start_message.database.report_type == FormulaReport:
-                pushers_formula[pusher_name] = supervisor.launch(pusher_cls, pusher_start_message)
+            if pusher.state.report_model == PowerReport:
+                supervisor.launch_actor(pusher)
+                print('Power Pusher Started')
+                power_pushers[pusher_name] = pusher
+            elif pusher.state.report_model == FormulaReport:
+                supervisor.launch_actor(pusher)
+                print('Formula Pusher Started')
+                formula_pushers[pusher_name] = pusher
             else:
                 raise InitializationException("Pusher parameters : Provide supported report type as model for pusher")
 
-        logging.info('CPU formula is %s' % ('DISABLED' if fconf['disable-cpu-formula'] else 'ENABLED'))
-        if not fconf['disable-cpu-formula']:
-            logging.info('CPU formula parameters: RAPL_REF=%s ERROR_THRESHOLD=%sW' % (fconf['cpu-rapl-ref-event'], fconf['cpu-error-threshold']))
-            setup_cpu_formula_actor(supervisor, fconf, route_table, report_filter, cpu_topology, pushers_formula, pushers_power)
+        logging.info('CPU formula is %s' % ('DISABLED' if global_configuration['disable-cpu-formula'] else 'ENABLED'))
+        if not global_configuration['disable-cpu-formula']:
+            print('creation of cpu dispatcher')
+            print('CPU formula parameters: RAPL_REF=%s ERROR_THRESHOLD=%sW' % (
+                global_configuration['cpu-rapl-ref-event'], global_configuration['cpu-error-threshold']))
+            logging.info('CPU formula parameters: RAPL_REF=%s ERROR_THRESHOLD=%sW' % (
+                global_configuration['cpu-rapl-ref-event'], global_configuration['cpu-error-threshold']))
+            setup_formula_actor(supervisor=supervisor, global_configuration=global_configuration,
+                                route_table=route_table, report_filter=report_filter, cpu_topology=cpu_topology,
+                                formula_pushers=formula_pushers, power_pushers=power_pushers,
+                                device_id='cpu', rapl_event=global_configuration['cpu-rapl-ref-event'],
+                                error_threshold=global_configuration['cpu-error-threshold'],
+                                dispatcher_name='cpu_dispatcher')
 
-            logging.info('DRAM formula is %s' % ('DISABLED' if fconf['disable-dram-formula'] else 'ENABLED'))
-        if not fconf['disable-dram-formula']:
-            logging.info('DRAM formula parameters: RAPL_REF=%s ERROR_THRESHOLD=%sW' % (fconf['dram-rapl-ref-event'], fconf['dram-error-threshold']))
-            setup_dram_formula_actor(supervisor, fconf, route_table, report_filter, cpu_topology, pushers_formula, pushers_power)
+            logging.info(
+                'DRAM formula is %s' % ('DISABLED' if global_configuration['disable-dram-formula'] else 'ENABLED'))
+        if not global_configuration['disable-dram-formula']:
+            print('creation of dram dispatcher')
+            logging.info('DRAM formula parameters: RAPL_REF=%s ERROR_THRESHOLD=%sW' % (
+                global_configuration['dram-rapl-ref-event'], global_configuration['dram-error-threshold']))
+            setup_formula_actor(supervisor=supervisor, global_configuration=global_configuration,
+                                route_table=route_table, report_filter=report_filter, cpu_topology=cpu_topology,
+                                formula_pushers=formula_pushers, power_pushers=power_pushers,
+                                device_id='dram', rapl_event=global_configuration['dram-rapl-ref-event'],
+                                error_threshold=global_configuration['dram-error-threshold'],
+                                dispatcher_name='dram_dispatcher')
 
-        pullers_info = PullerGenerator(report_filter, report_modifier_list).generate(args)
+        pullers_info = PullerGenerator(report_filter, report_modifier_list).generate(global_configuration)
         for puller_name in pullers_info:
-            puller_cls, puller_start_message = pullers_info[puller_name]
-            supervisor.launch(puller_cls, puller_start_message)
+            puller = pullers_info[puller_name]
+            supervisor.launch_actor(puller)
     except InitializationException as exn:
         logging.error('Actor initialization error: ' + exn.msg)
-        supervisor.shutdown()
+        supervisor.kill_actors()
         sys.exit(-1)
     except PowerAPIException as exp:
-        supervisor.shutdown()
-        logging.error("PowerException Error error: %s", exp)
+        logging.error("PowerException error: %s", exp)
+        supervisor.kill_actors()
         sys.exit(-1)
 
     logging.info('SmartWatts is now running...')
-    supervisor.monitor()
-    logging.info('SmartWatts is shutting down...')
+    return supervisor
 
 
-class SmartwattsConfigValidator(ConfigValidator):
+class SmartWattsConfigValidator(ConfigValidator):
     """
     Class used that check the config extracted and verify it conforms to constraints
     """
+
     @staticmethod
     def validate(config: Dict):
         if not ConfigValidator.validate(config):
@@ -276,7 +368,7 @@ def get_config():
 if __name__ == "__main__":
 
     conf = get_config()
-    if not SmartwattsConfigValidator.validate(conf):
+    if not SmartWattsConfigValidator.validate(conf):
         sys.exit(-1)
     logging.basicConfig(level=logging.WARNING if conf['verbose'] else logging.INFO)
     logging.captureWarnings(True)
